Update react-native-passbase.podspec
Xcode 12 can sometimes fail to build without a module to depend on React-Core directly hence this change is necessary for all native modules on iOS. For more details please check: facebook/react-native#29633 (comment)

I had this issue and it took me a while to fix it. This was my fix and seems to be a necessary change per the comment linked above. Tested this on my own device locally and functionality doesn't change but the problem is fixed.